Ajustes na TopBar -Alterações de icones e reposicionamento de componentes -Consertado o bug na ateração do icone no modo tela cheia e normal
