+++ PrimeDocs Desktop/PrimeDocs Desktop.csproj
@@ -18,6 +18,8 @@
     <None Remove="Resources\Icons\InitialWindow\MD.png" />
     <None Remove="Resources\Icons\InitialWindow\minimize.png" />
     <None Remove="Resources\Icons\InitialWindow\minus.png" />
+    <None Remove="Resources\Icons\InitialWindow\PrimeDocs.png" />
+    <None Remove="Resources\Icons\InitialWindow\PrimeDocsBar.png" />
     <None Remove="Resources\Icons\InitialWindow\profile.png" />
     <None Remove="Resources\Icons\InitialWindow\TXT.png" />
     <None Remove="Resources\Icons\PrimeDocs.ico" />
@@ -38,6 +40,8 @@
     <Resource Include="Resources\Icons\InitialWindow\MD.png" />
     <Resource Include="Resources\Icons\InitialWindow\minimize.png" />
     <Resource Include="Resources\Icons\InitialWindow\minus.png" />
+    <Resource Include="Resources\Icons\InitialWindow\PrimeDocs.png" />
+    <Resource Include="Resources\Icons\InitialWindow\PrimeDocsBar.png" />
     <Resource Include="Resources\Icons\InitialWindow\profile.png" />
     <Resource Include="Resources\Icons\InitialWindow\TXT.png" />
     <Resource Include="Resources\Icons\PrimeDocs.ico" />
