--- PrimeDocs Desktop/TopBarPrimeDocs.xaml
@@ -1,14 +1,0 @@
-﻿<UserControl x:Class="PrimeDocs_Desktop.TopBarPrimeDocs"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:PrimeDocs_Desktop"
-             mc:Ignorable="d">
-    <Grid Background="#FF0003DC" d:Height="25" VerticalAlignment="Top">
-        <local:Profile HorizontalAlignment="Right" Margin="0,0,175,0" VerticalAlignment="Top"/>
-        <local:Config HorizontalAlignment="Right" Margin="0,0,150,0" VerticalAlignment="Top" Height="25" Width="25"/>
-        <local:WindowControl HorizontalAlignment="Right" Margin="0,0,0,0" VerticalAlignment="Top"/>
-
-    </Grid>
-</UserControl>
