+++ PrimeDocs Desktop/UI/Components/TopBar/Profile.xaml
@@ -6,8 +6,8 @@
              xmlns:local="clr-namespace:PrimeDocs_Desktop"
              mc:Ignorable="d">
     <Grid>
-        <Image d:Height="30"  d:Width="30" HorizontalAlignment="Right"  Source="/Resources/Icons/InitialWindow/profile.png" VerticalAlignment="Center" />
-        <TextBlock TextWrapping="Wrap" Text="Nome" FontFamily="Open Sans" VerticalAlignment="Center" HorizontalAlignment="Right" Margin="0,0,50,0" FontSize="25" Foreground="White"/>
-        <Button Focusable="False" Style="{StaticResource ProfileConfigButtonStyle}" d:Height="30" d:Width="250" BorderBrush="{x:Null}" Click="btInitialWindowProfile_Click"/>
+        <Image Height="30"  Width="30" HorizontalAlignment="Right"  Source="/Resources/Icons/InitialWindow/profile.png" VerticalAlignment="Center" />
+        <TextBlock TextWrapping="Wrap" Text="Nome" FontFamily="Open Sans" VerticalAlignment="Center" HorizontalAlignment="Right" Margin="0,0,50,0" FontSize="20" Foreground="White"/>
+        <Button Focusable="False" Style="{StaticResource ProfileConfigButtonStyle}" Height="30" d:Width="250" BorderBrush="{x:Null}" Click="btInitialWindowProfile_Click"/>
     </Grid>
 </UserControl>
