+++ PrimeDocs Desktop/UI/Components/TopBar/TopBarPrimeDocs.xaml
@@ -1,0 +1,20 @@
+﻿<UserControl x:Class="PrimeDocs_Desktop.TopBarPrimeDocs"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:PrimeDocs_Desktop"
+             mc:Ignorable="d"
+             d:DesignWidth="800"
+             d:DesignHeight="35"
+             >
+    <Border Background="#001BC9" CornerRadius="7" ClipToBounds="True">
+        <Grid Background="Transparent" Height="35" VerticalAlignment="Top" ClipToBounds="True">
+            <TextBlock Name="tbTopBarTitle" TextWrapping="Wrap" Text="Title" FontFamily="Open Sans" VerticalAlignment="Center" HorizontalAlignment="Left" Margin="37,0,0,0" FontSize="20" Foreground="White"/>
+            <local:UserMenu d:BorderBrush="Red" d:BorderThickness="2" HorizontalAlignment="Right" Margin="286,0,160,0" Height="33" VerticalAlignment="Top"/>
+            <local:WindowControl x:Name="windowControlTopBar" d:BorderBrush="Red" d:BorderThickness="2" Margin="630,4,10,0" VerticalAlignment="Top" HorizontalAlignment="Right"/>
+            <Image Margin="10,5,0,5" Source="/Resources/Icons/InitialWindow/PrimeDocsBar.png" Stretch="Fill" Width="17" RenderTransformOrigin="-1.844,-8.906" HorizontalAlignment="Left" Height="25"/>
+        </Grid>
+    </Border>
+
+</UserControl>
