+++ PrimeDocs Desktop/UI/Components/TopBar/UserMenu.xaml
@@ -1,0 +1,12 @@
+﻿<UserControl x:Class="PrimeDocs_Desktop.UserMenu"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:PrimeDocs_Desktop"
+             mc:Ignorable="d" d:Height="30" d:Width="300">
+    <Grid HorizontalAlignment="Center" Width="340" Height="30">
+        <local:Profile d:BorderBrush="Yellow" d:BorderThickness="2" VerticalAlignment="Center" HorizontalAlignment="Left" Width="250" Height="30" Margin="23,0,0,0"/>
+        <local:Config d:BorderBrush="Yellow" d:BorderThickness="2" VerticalAlignment="Center" Height="30" Margin="290,0,20,0"/>
+    </Grid>
+</UserControl>
